lvrt-crash-dumps: document new LVRT core dump workflow
NILRT 9.1 / LVRT 23.1 changes the workflow to enable core dumps on LVRT.
LVRT's initscript now checks the value of the [LVRT]EnableCoreDumps INI
token to determine if it should follow the system settings for core
dumps - which on x64 is now `unlimited`.

Add documentation for this new workflow to the "crash dumps" tutorial.

Signed-off-by: Alex Stewart <alex.stewart@ni.com>