workflows: insert CNAME designation when publishing
The canonical nilrt-docs.ni.com subdomain tries to redirect to GH pages.
And it requires the Github custom domain configuration subsystem to
work, which means that the gh-pages ref must have a CNAME file with the
correct CNAME in it. But since that branch contents is entirely composed
of the built html files, we must insert the CNAME file in the workflow.

Signed-off-by: Alex Stewart <alex.stewart@emerson.com>